Prototype rolling own git history walking
When opening a large repository (e.g. the linux kernel), if we want to
sort by time + topology libgit2 takes a long time before returning the
first commit. Before we even read the first id on my machine I'm at
~12s. This is a long time and I'm fairly sure I can do better!

Prototype an implementation that walks the revision list and sorts it
correctly in less time. Leverage the knowledge that we do not need to
unzip entire commits, just their headers. This results in some pretty
ugly code that makes some pretty heavy assumptions, but it seems to be
good enough to parse a few repos for now. Eventually the plan is to hook
this up to the GUI and fall back on libgit2 if/when my parsing ever
fails.

This implementation on my machine is around 50% faster. I can walk the
linux tree in 6s instead of 12. This still isn't great but is somewhere
to start. There are still some wasteful allocations and copies that we
could avoid. For most repos I'll be working with this should be good
enough for a fairly instant response when I change which branches/tags
I'm looking at

Code is not very good quality. Some cleanup is still in order.