Push git2 down into wrapper
This gives our API a little bit more of a clear definition. We can
handle all the libgit2 switching internally. This will make it easier to
use our own implementations of features when we want to in the future

While we're at it move ObjectId into it's own file. It's gained enough
content that it's cluttering up git/mod.rs