Implement lazy fetching of commit data
Workflow is now as follows

* Lookup the metadata for each commit so we can build a graph
* Send the graph to the UI
* The UI tries to render every commit, if it doesn't have the full
  commit yet it just renders the commit id
* For each commit that we attempted to render, request the full commit
  from the app
* Cache the last 1000 commits

I thought that this would be pretty slow, but on my machine with an SSD
I almost never see commit hashes. Opening the linux repo and scrolling
to the bottom feels snappy once the repo is opened.

For now this is good enough. I expected to have to do some paging or
something to pre-load pages, but we'll save that for a later day.