Initial branch filtering implementation
Initial implementation of branch filtering. This implementation still
has known problems. On large repositories it isn't fast enough and
there's no indication that work is going on behind the scenes. If you
spam click multiple branches on a large repository, a lot of extra work
is done since we re-build the commit graph on every interaction. This
initial implementation is just to get something on the screen, future
revisions can make this a little more robust.

* Push available branches to the GUI
* Store selected branches in the GUI and update the app when they change
* Add graph building filters
* Fix metadata iter using all loaded commits instead of only walked
  commits
* Add side panel with branch checkboxes
* Fix UI not updating when branch finishes loading
  * Process events in their own thread and request a UI update when a
    new event comes in