Add feature to download arbitrary remote references
A common workflow for me is to open a PR on whatever git service I'm
using, then look for a reference that matches that given PR. The naming
of these references is different depending on the service being used.

Add a searchable window of remote references

* Add repo function to list remote references. This is reasonably slow
  and can block up the app. For now I think that this is manageable, but
  if slow downs are noticed an async fetch may be necessary
* Add toolbar button to open new dialog
* Add command line to download a remote reference as <remote>/refs/...
* Remotes are only fetched when the dialog is open. I experimented with
  fetching on repo open, but that noticeably slowed down the time to
  open a repo.

Currently there is no way to remove those references from the UI which
is a little annoying. Implementing a deleter for remote branches is
unclear if it will delete the remote reference, or the local pointer to
the remote reference.