Allow sorting based off either author or committer timestamps
Sometimes both are useful. It's useful to see author timestamps when
comparing to old rebased commits, and it's useful to see committer
timestamps to see how the repo was modified over time. Allow sorting
based off either

* Default to committer timestamps as they're generally more in line with
  how I would want the repository to be presented
* Update commit parsing logic to continue parsing until committer
  timestamp is found
  * Bonus cleanup in the decompress area since this is very similar to
    how looking for the author works
* Add dropdown in GUI for selecting which sorting method we want to use
* Add sorting option to the view state request for the graph
* Add sorting option to metadata index sorting

Tests are light for no good reason, but this parses successfully on the
linux repo and sorting looks correct when I change the sorting method on
some PRs that have been rebased a few times in a couple repos