Initial tag support
Initial tag support. Note that this does not have any support for
viewing tag annotations, or interacting with the tag object itself. At
this point we view the tag in the same way we view a symbolic reference
or a branch. This allows us to avoid implementing tag parsing in our
packfile implementation.

* Rename branch -> reference
* Add tags as a reference type