[WIP]: Interactive command line runner
Current state is that we have a command line runner which _should_
report back to the gui when there is output and accept input from the
gui to stdin. However it seems that there is something about unix pipes
I do not understand.

If I do `git add -p | tee`, I do not get any output until I spam ? a few
times. This persists even if I use setbuf to set the input and output
buffer lengths to 0. This also results in my application not handling
this case well