Improve libgit2 metadata fallback
Initial libgit2 fallback was done at too high of a level. Really we
should be able to just fallback to libgit2 for specific objects we are
unable to get, but fill our metadata storage the same way.

This has a performance cost on populating our metadata storage, as if
there are a lot of ref deltas we will keep flip flopping between our way
and libgit2's way, but it should speed up behaviour when we change
branches after doing the initial load of a repo